fix: format date ranges in tooltip component
This fixes a bug where the tooltip show date ranges in different formats
when changing the range and updating the visualization.
